Fix isOwner guard for edit and delete

+++ SWE-academy/SoftUni-Back-End-Sept-2025/14-Exam-Preparation/src/services/blogService.js
@@ -25,11 +25,18 @@ function follow(blogId, userId) {
     return Blog.findByIdAndUpdate(blogId, { $push: { followers: userId } });
 }
 
-function remove(blogId, userId) {
+async function remove(blogId, userId) {
+    const blog = await Blog.findById(blogId);
+
+    if (!blog.owner.equals(userId)) {
+        throw new Error('Cannot delete if not owner');
+    }
+
     return Blog.findByIdAndDelete(blogId);
 }
 
 function edit(blogId, blogData) {
+
     return Blog.findByIdAndUpdate(blogId, blogData, { runValidators: true });
 }
